requirements.txt
cycler==0.10.0
Django==1.10.5
functools32==3.2.3.post2
image==1.5.5
matplotlib==2.0.0
numpy==1.12.0
olefile==0.44
opencv-python==3.2.0.6
Pillow==4.0.0
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
scipy==0.18.1
six==1.10.0
subprocess32==3.2.7
